.r D6: extract GitHub Actions workflow format from pinnedinputs.go (behavior-identical; tests pass unchanged)
Split the workflow pinned-input format out of the 1,743-line
internal/metadatautil/pinnedinputs.go into a focused sibling file,
pinnedinputs_workflows.go, per roadmap item D6 (split oversized Go
validators per format).

Pure relocation within package metadatautil — no logic changes. Moved the
standalone, format-cohesive declarations: the workflow regex vars
(workflowPermissionsRE, actionRefPattern, commitShaPattern), the uses-scan
types + extractWorkflowUses, toolPins/loadToolPins/parseToolPins,
loadAllowedActions, IsSafePullRequestTargetWorkflow, and the workflow YAML
helpers (parseWorkflowRoot, yamlMappingValues, yamlScalarValue, mustGlob,
workflowYAMLFiles). Dropped the now-unused sort and yaml.v3 imports from
pinnedinputs.go; the new file carries its own imports.

Exported API unchanged: PinnedInputsConfig and CheckPinnedInputs stay in
pinnedinputs.go; IsSafePullRequestTargetWorkflow stays exported from the new
file. Callers and test files are untouched (byte-identical test logic).

The CheckPinnedInputs orchestrator (docker/node/rust/python format checks) is
inlined as shared closures with interleaved cross-format validation, so it
cannot be split by format as a pure relocation — that requires a follow-up
refactor promoting the shared closures to package functions first.

Verified: go build ./..., go vet, gofmt -l clean; go test ./... green; the
metadatautil workflow/pinnedinputs tests pass unchanged.
